feat: Implement circular data structures for efficient FIFO storage
- Added CircularQueueNode class for managing circular queue nodes.
- Introduced SharedInt32CircularBuffer for shared memory circular buffer operations.
- Created CircularBuffer class for fixed-size FIFO storage with optimized read/write.
- Developed UnrolledQueue class utilizing CircularQueueNode for dynamic queue management.
- Implemented UnrolledTaskQueue for high-performance task processing with batch capabilities.
- Added index file to export all data structure modules for easier imports.
