+++ src/lib/core/public/data-structures/index.ts
@@ -1,6 +1,1 @@
-export { default as CircularBuffer } from "./Buffers/CircularBuffer";
-export { default as SharedInt32CircularBuffer } from "./Buffers/SharedInt32CircularBuffer";
-
-export { default as createListenersSet } from "./ListenersSet";
-export { default as UnrolledQueue } from "./UnrolledQueue";
-export { default as UnrolledTaskQueue } from "./UnrolledTaskQueue";
+export * from "./public";
