+++ src/lib/core/public/data-structures/public/index.ts
@@ -1,0 +1,3 @@
+export { default as CircularBuffer } from "./CircularBuffer";
+export { default as UnrolledQueue } from "./UnrolledQueue";
+export { default as UnrolledTaskQueue } from "./UnrolledTaskQueue";
