+++ src/lib/core/public/os/index.ts
@@ -2,4 +2,3 @@ export * from './support';
 export * from './browser';
 export * from './platform';
 export * from './controls';
-export * from './env';
